fix: 主干告警清零 【背景】 主干中存在告警需要处理
【修改方案】
在有符号数与无符号数混用的地方使用了强制类型转换

【影响】
对现有的产品编译不会有影响。

re #I4SPC0

+++ kal/posix/src/time.c
@@ -251,7 +251,7 @@ STATIC VOID OsGetHwTime(struct timespec *hwTime)
                      (cycle % OS_SYS_CLOCK) * OS_SYS_NS_PER_SECOND / OS_SYS_CLOCK;
 
     hwTime->tv_sec = (time_t)(nowNsec / OS_SYS_NS_PER_SECOND);
-    hwTime->tv_nsec = nowNsec % OS_SYS_NS_PER_SECOND;
+    hwTime->tv_nsec = (long)(nowNsec % OS_SYS_NS_PER_SECOND);
 }
 
 STATIC VOID OsGetRealTime(struct timespec *realTime)
